requirements.txt
amqp==5.1.1
asgiref==3.6.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.0.3
Django==4.2
django-auditlog==2.2.1
django-cors-headers==3.13.0
django-jazzmin==2.6.0
django-mysql==3.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
mysqlclient==2.1.1
packaging==22.0
pika==1.3.1
Pillow==9.4.0
prompt-toolkit==3.0.38
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python-http-client==3.3.7
pytz==2023.3
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sendgrid==6.9.7
six==1.16.0
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.6
